[Fix] Fixed type-casting in ft_substr #4
In line 24, my previous code was casted the result of strlen(s) and
compare the result to start(unsigned int).

In that situation, if length of s is more than maximum value of unsigned
int, program is not going to work properly.

I fixed the situation, casting start to size_t.